feat(data): 입력 검증, BigDecimal 변환, 입찰 비관적 락 + 동시성 테스트

+++ src/main/java/com/akiba/backend/auction/dto/response/AuctionPostListResponse.java
@@ -6,8 +6,11 @@
 package com.akiba.backend.auction.dto.response;
 
 import lombok.Builder;
+import java.math.BigDecimal;
 import lombok.Getter;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @Getter
 @Builder
@@ -16,9 +19,9 @@ public class AuctionPostListResponse {
     private Long postId;
     private String title;
     private String specialType;
-    private int startPrice;            // 시작가
-    private Integer currentPrice;      // 현재 최고 입찰가
-    private Integer buyNowPrice;       // 즉시구매가
+    private BigDecimal startPrice;            // 시작가
+    private BigDecimal currentPrice;      // 현재 최고 입찰가
+    private BigDecimal buyNowPrice;       // 즉시구매가
     private int bidCount;              // 입찰 수
     private String thumbnailUrl;
     private LocalDateTime endsAt;      // 경매 종료 시간
