+++ src/main/java/com/akiba/backend/market/dto/response/MarketPostListResponse.java
@@ -5,8 +5,11 @@
 package com.akiba.backend.market.dto.response;
 
 import lombok.Builder;
+import java.math.BigDecimal;
 import lombok.Getter;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @Getter
 @Builder
@@ -15,7 +18,7 @@ public class MarketPostListResponse {
     private Long postId;
     private String type;              // USED, LIMITED
     private String title;
-    private Integer price;
+    private BigDecimal price;
     private String productCondition;
     private String specialType;
     private String status;
